gist-1 clone command - refactor (move context to context.go).
